Make the PhishingController `test` synchronous (#929)
The `test` method for the PhishingController was recently made
asynchronous so that it could update the phishing configuration if
necessary. We found that in practice this was difficult to use,
especially when handling multiple simultanous tests.

The `test` method has been made synchronous again. Instead we can check
whether the config is out of date using the new `isOutOfDate` method,
and explicitly update the configuration if required.