Handling the error when eth_estimateGas fails (#920)
When we did a transaction with binance smart chain and the dapp doesn't send a gas value, we were not being able to execute the eth_estimateGas, it was falling do a general RPC error.

Handle the error like the extension does it. Added a value to the Transaction object to be filled with an error if the eth_estimateGas fails.