Fix occasional type errors when running tests (#935)
I have noticed that locally `yarn test` will succeed the first time,
the will fail with a type error saying `toNeverResolve` is undefined.
After running `yarn jest --clearCache`, tests will pass again once then
fail again.

This custom Jest matcher was added in #930. Seemingly the way the type
was added is causing this problem.

I'm not entirely sure what's going on here. Something about the way
Jest is loading this file is making Jest confused about this type.
I found that it works more reliably if the type is moved to a separate
file in the `types` directory.