Remove outdated sections from README
The "usage" and "concepts" section of the README were out of date, so
they have been removed. Portions of the copy have been migrated into
JSDoc comments, but most of it I couldn't find a home for.

We will soon be dedicating time to write thorough usage documentation
to explain the patterns that we are using and recommending today.

The section on linking has also been removed because it was specific to
npm. We still use a workflow similar to what was described, but it
wasn't close enough to justify leaving it. We can try to replace it
soon as we're writing usage documentation. It's likely to change anyway
as we continue migrating to Yarn v3, and using `portal:` over `link:`.