Assign a return type to setTimeout that works in both Node and browser envs
In Node, `setTimeout` returns NodeJS.Timer, whereas in the browser,
`setTimeout` returns a number. So we if assign a type of NodeJS.Timer to
the variable here (see diff), `tsserver` will be cool with it, `tsc`
will be cool with it, but Jest will not be cool with it (because we've
instructed the `transaction-controller` tests to run within a jsdom
environment, and so `@types/node` [somehow] does not apply in this
case).

Also ensure that when we use the Jest `jsdom` environment we clear out
the existing options we'd set for the `node` environment, just so we
know that Jest isn't trying to do something we don't expect it to.