Remove `is-release` condition (#981)
* Remove `is-release` condition

The `is-release` job was configured to only run for commits made by
GitHub actions. Our old release workflow always started with a commit
made by an action, so this was an easy way to skip all release steps
for commits that obviously were not releases.

Our new workflow does not work this way, so this check no longer makes
sense.

The `action-is-release` action should still work in determining whether
or not the current commit is a release. It checks the contents of the
commit, specifically whether or not the base manifest version has
changed. This technique still works for our new workflow.

* Only run is-release for push events to the main branch