Add documentation
Those of us that have worked on the monorepo have a lot of context
around why it exists, what it does, and how to work with it, but
for everyone else, this context may not exist. The set of guides in this
commit aims to fill in the missing gaps. I've also included a bit around
preview builds at the end. We may also need an expanded section on
creating new releases, as people may not immediately jump to the docs
for `create-release-branch`, and those docs may not be as comprehensive
as they could be. This could come in a future PR, however.