Add peer dependencies between packages
Packages now list `peerDependencies` to represent the expectation that
a specific package version is setup with the controller messenger. This
ensures that projects using these packages get a warning upon install
if they are using incompatible package versions.

This lets us safely make breaking changes to controller messenger
events and actions.

The constraints needed adjustment to allow the same dependency to be
declared both in `dependencies` and `peerDependencies`. The constraint
now ensures packages aren't listed both in `dependencies` and
`devDependencies` instead.