Align Provider type w/ other packages & re-expose (#117)
The Provider type which is used internally by this package is not
compatible with the SafeEventEmitterProvider type used by
`eth-json-rpc-middleware`. This means that whenever using this package
with `eth-json-rpc-middleware` we have to resort to the following
hackery:

```
import { PollingBlockTracker, Provider } from 'eth-block-tracker';
import { JsonRpcEngine } from 'json-rpc-engine';
import { providerFromEngine } from '../src';

const engine = new JsonRpcEngine();
const provider = providerFromEngine(engine);

const blockTracker = new PollingBlockTracker({
  // Notice the typecasting here
  provider: provider as Provider,
});
```

SafeEventEmitterProvider is the more accurate type here in this case,
because it correctly types the `error` argument for the callback that
`sendAsync` calls as `unknown` rather than `Error`, so we follow suit
here.

Additionally, this package used to expose the Provider type and that was
changed in 6.0.0 without being documented, probably because it was
located in the same file as BaseBlockTracker. So here we move the type
to a more obvious place and re-expose it.