Allow network requests in instances of the CurrencyRate and TokenRate controllers to be disabled (#1002)
**Allow network requests in instances of the CurrencyRate and TokenRate
controllers to be disabled**

**Description**

- BREAKING:
- A new private property is added to the CurrencyRateController:
`enabled`. If this is false, then the function in that controller which
makes network requests will not make any requests when called. This
property is controlled by the `start` and `stop` methods, so that these
methods will effectively enable or disable network requests. Previously,
even if `stop` had been called, `setNativeCurrency` or
`setCurrentCurrency` would trigger a network request. That is now
prevented.

- FIXED:
- The TokenRatesController previously had a bug whereby if a `disabled`
property was passed to the constructive as part of the `config` object,
it would be ignored / overwritten to false. This PR fixes that.


**Checklist**

- [ ] Tests are included if applicable
- [ ] Any added code is fully documented

**Issue**
Part of
[#16724](https://github.com/MetaMask/metamask-extension/issues/16724)

Co-authored-by: Dan J Miller <danjm.com@gmail.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: Pedro Figueiredo <pedro.figueiredo@consensys.net>
Co-authored-by: Pedro Figueiredo <ganseki.figueiredo@gmail.com>