Optionally move addresses when RPC chain ID changed
## Context

The AddressBookController maps addresses to nicknames. As an address is
chain-specific, there are multiple mappings, one per chain id. While a
chain id seems global, it can actually be traced back to a network that
is visible within a given product under the user settings panel.

In the extension, when the user updates the chain id for a network, then
the mapping for that chain id is migrated from the previous chain id to
the new chain id. This logic for performing this migration is currently
kept within the PreferencesController along with code to manage the
network list itself. This is undesirable for two reasons:

1. We want to move network management code out of PreferencesController
to NetworkController. This includes the migration logic.
2. One could argue that it is not PreferencesController's responsibility
to keep address book state in sync with the current list of networks.

Finally, it is worth noting that no such migration behavior exists for
mobile.

## Changes

Given this context, this commit copies the migration logic from the
extension PreferencesController to AddressBookController in this repo.

Note that as this controller is shared between both extension and
mobile, the logic is disabled by default so that the extension and
mobile app can operate as they do now. A consumer that wishes to use
this feature must pass three new options to AddressBookController:

* `syncWithRpcChanges`: must be true.
* `onPreferencesStateChange`: a function which will be called when the
PreferencesController state changes; you are expected to wire this up
appropriately (see other examples in MetamaskController)
* `getProvider`: a function which returns the MetaMask provider
(obtained from NetworkController)

None of these options is required.