Use GH Actions for parallelism rather than Yarn (#1029)
`yarn test` is a multi-package command: it really runs `yarn test` for
each package in parallel using `yarn workspaces foreach`. This works
fine as long as all of the tests pass. But because this command produces
a lot of output, if tests fail for any package, it can be very difficult
to locate the failure. (The same goes for `yarn validate:changelog`.)

There are a few ways we could solve this when these commands are run
locally, but it's particularly a pain point when run on CI. For this
reason, this commit adjusts the GitHub Actions configuration so that we
use GitHub Actions instead of Yarn to run these commands in parallel
across the monorepo. This means that if tests fail for a particular
package, you only see output for that package and no other package.