Update Yarn constraints to remove reference to project/repo name (#1032)
When creating the monorepo template, it was useful to revise the Yarn
constraints file so that the project/repo name could be changed without
having to change a bunch of references in the file. This commit brings
over those revisions:

* The repository URL is now used as a basis for the homepage URL and
bugs URL, so it must be set appropriately (and if it is not, the user
will see a suggestion in the correct format).
* Additionally, the comments above each constraint have been reworded to
be more consistent and understandable, and comments have been added
where missing.
* Finally, some pairs of constraints which test one thing for private
packages and the opposite for public packages have been updated to apply
to the root package vs. non-root packages where it made sense.