Rewrite NetworkController unit tests
There are a few issues with the existing unit tests for
NetworkController:

* They aren't as comprehensive as they ought to be. While there are
tests for some essential behavior that the controller provides, there
are numerous code paths that the controller takes as it creates and
interacts with a provider, and the extent of these paths is not fully
tested.
* Also, it is difficult to tell what behavior is exactly is being
executed, as tests are flatly organized, and it would be more manageable
if tests were organized by method name.
* Finally, there are some tests which ends up running network requests
or out-of-band asynchronous code, and this must be properly stubbed.

Unfortunately, to properly address these issues, the unit tests needed
to be completely rewritten. In the course of doing this, type
definitions were added for `web3-provider-engine`, as this is helpful to
ensure that mocking `web3-provider-engine` behavior in tests is done
correctly.

Note that rewriting the tests uncovered several quirks with the existing
behavior. For instance:

* `lookupNetwork` makes an asynchronous network request, but is not an
asynchronous method itself. This means that some code could instantiate
a controller, use the controller to initialize a provider, then discard
the controller, and there is a chance that the request may complete out
of band. This actually happens in the tests because they run so fast,
so this is the only quirk I've gone ahead and fixed. To do this I've
added a `destroy` method to NetworkController and updated
`lookupNetwork` to do nothing if `destroy` has been called.
* `lookupNetwork` publishes an `NetworkController:providerConfigChange`
event via the controller messenger, even though this method never
changes the provider config.
* Based on the API, it appears that the provider config can be specified
two ways: ahead of time via the constructor and on the fly via the
`providerConfig` setter. Despite their apparent similarity, these
provider configs are kept separate and are actually used for two
different purposes. The provider config set via the constructor is
stored in state and is to determine which kind of provider to build
(Infura vs. custom RPC), whereas the on-the-fly provider config is used
to actually customize the provider. This is likely to be very confusing
for users.
* If the `providerConfig` setter is used and the `type` within the
in-state provider config is `rpc`, but no `rpcTarget` property is
present, then the setter will do nothing.
* `registerProvider` ends up getting called twice when the
`providerConfig` setter is used. This happens because
`initializeProvider` is called, then either `setupInfuraProvider` or
`setupStandardProvider`, then `updateProvider`, then `registerProvider`;
however, the `providerConfig` setter calls `registerProvider` itself.
Because `registerProvider` is responsible for responding to the "error"
event on the provider, when that event occurs then `lookupNetwork` will
get called twice. In practice this isn't an issue because of the mutex
in `lookupNetwork` but it does lead to more network requests than
necessary.
* Calling `setProviderType` with a type of "rpc" doesn't end up setting
a provider. This method is probably not meant to be used this way — it's
probably only designed to take Infura network names.