Expose prev state in state change event payloads
When listening for controller state changes, it is useful to be able to
tell whether an individual property in state changed. However, this is
not possible currently because the state change event only exposes the
state _after_ it has been changed, not _before_. This commit adds a
third argument to the `stateChange` listener so that the previous state
can be accessed.