Refactor function to get cache identifier (#182)
* Refactor function to get cache identifier

The function `cacheIdentifierForPayload` has been renamed to
`cacheIdentifierForRequest`, and refactored to accept a JSON-RPC
request rather than the `Payload` type.

The function `paramsWithoutBlockTag` has been updated to accept a JSON-
RPC request rather than a `Payload` as well. I noticed that this
function wasn't being used anywhere, so the `export` was removed as
well.

The `block-cache` and `inflight-cache` middleware required some
internal updates to accomodate these type changes. No functional
changes or external-facing type changes have been made.

Documentation and tests have been added. This brings test coverage for
`cache.ts` up to 100%.

Relates to #176 in that it helps unblock the necessary changes.

* Make `method` parameter mandatory for cache utility functions

The `method` parameter is now mandatory for the cache utility
functions. It was only made optional in the first place as a side-
effect of an erroneous type that was being used (`Payload`), which has
now been replaced.

* Remove leftover merge marger