Migrate both provider functions (#6)
* Migrate both provider functions

The functions `providerFromEngine` and `providerFromMiddleware` have
been migrated from `@metamask/eth-json-rpc-middleware`.

The `SafeEventEmitterProvider` is now a class with its own separate
module, rather than a type. Hopefully this will help prevent future
type errors. The error in the `SafeEventEmitterProvider` type in
`@metamask/eth-json-rpc-middleware` has been corrected; the response in
in the callback is now optional.

Everything should be fully documented and tested.

* Export type

* Fix spelling mistake

Co-authored-by: Alex Donesky <alex.donesky@consensys.net>

* Improve test names

---------

Co-authored-by: Alex Donesky <alex.donesky@consensys.net>