Improve compatibility of `hasProperty` (#80)
* Improve compatibility of `hasProperty`

The `hasProperty` function is now compatible with a wider variety of
objects. The `Record<PropertyKey, unknown>` constraint used previously
was exluding `Error`s, and any object-like things expressed as a custom
class or interface. Now all of those are accepted.

* Ignore `tsd` tests when building

These test modules are not meant to be compiled directly with
TypeScript. They sometimes include errors by-design that would blow up
if building with TypeScript.

* Re-enable TypeScript in `test-d.ts` files to enable linting

The type error assertion was replaced with an alternative that does not
introduce a compile-time error.