Replace network/isCustomNetwork with networkStatus
The network controller contains two related state properties: `network`
and `isCustomNetwork`.

The first property is strange: it can either be a number, which
represents the version of the network as obtained via `net_version`, or
"loading", which is a sort of null state that either means we don't have
a network to connect to, or we do but it's not accessible. In any case,
the `net_version` RPC method, much less the idea of a network version,
is no longer standard, so we don't need to store it.

The second property is unnecessary, as we can already obtain whether a
network is "custom" (i.e. not one of the ones we know about by default)
by looking at the `type` within the provider config object.

Therefore, we can remove these state properties and replace them with a
simple property that merely tracks whether we have an accessible network
or not.

Co-authored-by: Salah <salah-eddine.saakoun@consensys.net>