Fix Yarn constraints to allow for github: refs
To test a new release of a library that lives in another repo, it is
common to commit the compiled version of that library in a branch and
push that branch, then come back to this repo and update `package.json`
to point to the repo/branch. The current set of Yarn constraints does
not allow `github:` references in dependency fields. This commit allows
for this. (Note that the constraint here is naive and expects you to
provide a valid repo reference.)