Fix/web3 middleware testing (#1087)
* fix: add network client tests from extension

* fix: update yarn lock

* revert network controller

* Fix stub for isEIP1559Compatible

* Fix lint configuration

The provider API test modules weren't being recognized as tests by
ESLint. The config has been updated to apply Jest lint rules to them.

A bunch of lint rules have been disabled for now on a per-file basis,
because the structure of the shared tests is incompatible with the
rules. We did something similar on the extension, except less rules
needed to be disabled there. We should give these a pass later to
consider whether we can re-enable any of them.

The only lint fixes made were to comply with the preference for
`toBeUndefined` over `toBe(undefined)`.

* Add methods not supported by Infura

This is a port of the changes from this PR: https://github.com/MetaMask/metamask-extension/pull/17347

* Add tests for `retryOnEmpty` middleware

This is a port of the tests added in this PR: https://github.com/MetaMask/metamask-extension/pull/17364

* Remove retry on empty tests

It seems that the network client does not retry on empty.

* fix: put logging behind the debug

* Fixed ethquery type alias

* Remove extraneous NockScope cast

* Fixed the typings for mockRpcCall

* Added exported type for ProviderType and drop related checks

* Fixed spies on networkcontroller to spy on instance instead

* Remove additional checks covered by typescript

* Remove extra checks and use imported ProviderType

* Fixed indentation issue for timeout retry test

* Fixed phrasing of tests with static responses

* Fixed additional indenting issue

* Fixed comments relating to block-ref not being present in web3providereng

* Fixed wording on empty response tests

* Fixed typo

* Fixed more typos

* Remove accidentally committed experiments

* Fixed json-schema-tools/meta-schema to be a devdep

* Moved tests into tests directory

* Fixed eslint config to lint test folders and ignore certain rules

* Removed extra comments

* Moved tests for methods not handled by middleware to their own file

* Fixed rebase issues

* Moved network client test helpers to their own files

* Fixed linting issues

* Fixed linting rule feedback

* Fixed tests that broke after rebase

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>