Generate source maps for type definition files
It is a frustrating experience to want to see the source code of a
module that's being imported in one's editor only to be presented with
the `.d.ts` file when the module is Cmd-clicked (or "Go to Definition"
is otherwise invoked). To fix this, TypeScript needs to be configured to
generate source maps alongside `.d.ts` files.