Make `verifyNetwork` async
The internal method `verifyNetwork` is now async. It performed an
async call but didn't `await` it.

This method is only invoked as an internal event handler, so this
didn't require any changes outside of that function (not even to
tests). Effectively this is not a functional change in any way.

This relates to https://github.com/MetaMask/core/issues/1176