Make `upsertNetworkConfiguration` async (#1192)
The network controller method `upsertNetworkConfiguration` is now
async. If the option `setActive` is false, the flow of execution should
be the same as before (apart from the change in return type). But if
`setActive` is `true`, the method will not resolve until the network
switch has completed (previously it would initiate the switch without
waiting for it to complete).

This relates to #1176