Replace `network` state with `networkId` and `networkStatus`
The network controller `network` state used to be set to `loading` if
the network was loading or uninitialized, or to the network ID if the
network had finished loading. Effectively it was tracking both the
network ID for the current selected network, and the network status.

This state property has been split in two; now we track the network ID
of the current selected network separately from the network status. The
network status has been expanded to include more states as well.

Closes #1020