Remove `isCustomNetwork` (#1199)
* Remove `isCustomNetwork`

The network controller state `isCustomNetwork` has been removed. This
state was redundant; you can determine whether the current selected
network is custom or not by checking the `providerConfig` state.

The one usage of this state in other packages was the transaction
controller. It has been updated to use the provider configuration state
instead.

Closes #1020

* Remove unused private method