Refresh network when connection is reset
The `resetConnction` method has been updated to match the extension
more closely. We now refresh the network when this is called, rather
than just reconstructing the provider and checking EIP1559 status.

Unit tests have been updated to align more closely with the extension
as well.

This was done to unblock #1203, but it also relates to #1023