Add handling of Infura "blocked" status
The network controller has been updated to detect when connecting to an
Infura network fails due to the user being blocked. A new status has
been added to reflect this state, and two events have been added (one
for when a user is blocked, and one for when they are unblocked).

This status and these events are used by the extension to better
explain why the network is not working in this scenario.

Fixes #1200