Add `eth-query` types
Type definitions have been added for the `eth-query` package. These
were copied from the type definitions used in the extension.

The addition of these types revealed various invalid assumptions we've
made throughout the codebase, requiring the addition of runtime checks.
Mostly this consists of throwing an error whenever someone tries to use
the `ethQuery` instance before it has been set. In most cases, this
would only introduce an error in code paths that would already throw an
error.

The one exception to that is the gas fee controller, which used to get
`ethQuery` upon construction. It has been updated to wait until the
`providerConfigChange` event before attempting to fetch that, since
it's never defined upon construction in practice anyway.

There are two places in the network controller where these types
revealed that we're not validating the network response to the
`net_version` and `eth_getBlockByNumber` calls made during network
lookup. A comment has been left about validating this in a future PR.

Closes #1204