Fix GasFeeController to assign ethQuery initially
GasFeeController uses the same instance of EthQuery that
NetworkController constructs when it constructs a new provider. It will
then use that EthQuery instance later in order to make requests to the
network.

A previous change removed a line from GasFeeController which accessed
that EthQuery instance immediately in the constructor because it was
thought that at point in time the provider had not been set yet.
However, even if do not `await` the call to `initializeProvider` that
occurs after NetworkController is instantiated, the provider will still
be constructed synchronously.

This commit restores the removed line and adds tests for it.