Extract network client construction from NetworkController (#1285)
The NetworkController will construct a provider that is designed to
interface with either an Infura network or a custom RPC endpoint,
depending on state, along with a block tracker. We call this pair the
network client, and in the extension we have extracted its construction
to a separate file. This commit follows the same pattern so that when we
replace `web3-provider-engine` we don't have to introduce so many
changes. It also consolidates the network client tests (which we still
call the provider API tests, though that will be renamed at some point).

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>