Remove warning in NetworkController unit tests
In the NetworkController unit tests we use `setTimeout` to wait for
events to occur. We also use `clearTimeout` in the same context to stop
waiting if need be. Because we use fake timers in these tests we need
the original versions of these functions and not the fake versions. We
already capture the original version of `setTimeout` up front, but we
don't do the same for `clearTimeout`, and that leads to a warning being
printed when we use it. This commit captures `clearTimeout` up front as
well to prevent this warning.