Add beforeCompleting option to FakeProviderStub (#1301)
In some NetworkController unit tests it is useful to mock a network
request, but then do something before the request completes, such as
switching the network while `net_version` is being requested. On the
extension side we've solved this by allowing a `beforeCompleting` option
to be specified when providing the network request stub. This commit
copies that same option.