Improve types used for network client test helpers
The types used for the network client test helpers have been made more
specific; the network type passed in is now required at a type-level to
be an Infura network.

This was a type-only change because we were already only passing in
Infura neworks for this option. This was done to make a later refactor
PR simpler.

Relates to #1116