NetworkController: Fix chain IDs in tests (#1307)
When we replace `web3-provider-engine`, we will intercept `eth_chainId`
to return a static result, which, for custom RPC endpoints, will come
from the chain ID in the provider config. This means that this chain ID
must be convertable to a 0x-prefixed hex string. However, when we
specify chain IDs in the NetworkController tests, we sometimes use
fake hex numbers. So this commit fixes various references in the tests
to fulfill this requirement and generally make chain IDs match what we
would expect them to be in reality.