NetworkController: Use the same middleware stack as the extension (#1116)
The network controller in this repo exposes a JSON-RPC provider object built via the `web3-provider-engine` package. However, we have not supported this package for a long time, and the network controller in the extension uses a more modern provider built from `json-rpc-engine` and `eth-json-rpc-middleware`. In order to unify the two versions of the network controllers, we need to replace the middleware stack in this version with the one in the extension.

Co-authored-by: Mark Stacey <markjstacey@gmail.com>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>