NetworkController: Remove providerConfigChange event (#1329)
Despite its name, the `providerConfigChange` event in NetworkController
doesn't get emitted whenever the `providerConfig` property in state is
changed. Rather, it gets updated as the very last step in
`lookupNetwork`. This method is called:

- When the provider is initialized
- When the network is switched
- When the network is rolled back to the previous setting
- When `resetConnection` is called

`lookupNetwork` can also be called on its own.

If this were not enough, we already provide a way to listen for changes
to a particular property in state: by passing a selector function to the
`subscribe` method on the controller messenger.

So, this commit removes the `providerConfigChange` event from
NetworkController. Instead of the following code:

```
messenger.subscribe(
  'NetworkController:providerConfigChange',
  (providerConfig) => {
    // do something with the providerConfig
  },
)
```

consumers should be able to say:

```
messenger.subscribe(
  'NetworkController:stateChange',
  (providerConfig) => {
    // do something with the providerConfig
  },
  (networkControllerState) => networkControllerState.providerConfig,
)
```