NetworkController: Match `rollbackToPreviousProvider` to extension (#1223)
When `rollbackToPreviousProvider` is called, it will use
`#previousNetworkSpecifier` to determine how to switch the network back
to the previous one. This property holds one of two values depending on
how that network was set in the first place:

- If `setActiveNetwork` was used to set the network, then
  `#previousNetworkSpecifier` will be the ID of a network configuration.
- If `setProviderType` was used to set the network, then
  `#previousNetworkSpecifier` will be the `type` in the provider config.
- If the network controller has just been initialized, then if the
  `type` in the provider config is "rpc" and the provider config holds a
  reference to a network configuration, then `#previousNetworkSpecifier`
  will be its ID, otherwise it will be whatever the provider `type` is.

This means that when `rollbackToPreviousProvider` is called, it can ask
what `#previousNetworkSpecifier` is and use its value to route to the
correct method:

- If `#previousNetworkSpecifier` is a known provider type, then call
  `setProviderType`, passing the type.
- Otherwise, if `#previousNetworkSpecifier` is a string, assume that it
  is a network configuration ID and call `setActiveNetwork` with it.

However, a problem arises when the network has not been switched yet,
and the provider config has a `type` of "rpc" but does not contain a
reference to a network configuration. Because
`#previousNetworkSpecifier` in that case is a provider type,
`rollbackToPreviousProvider` would try to call `setProviderType` with
"rpc". `setProviderType` is meant to be passed an _Infura_-specific
provider type, i.e. the name of an Infura network, so it will throw if
it doesn't receive one.

There are probably a couple of different ways to solve this problem,
but we can consider what the extension version of NetworkController
does: when the controller is initialized or before the network is
switched, the current provider config is saved, the whole object is
loaded into `providerConfig` in state when `rollbackToPreviousProvider`
is called, and then `setProviderType` and `setActiveNetwork` are
completely skipped.

Because that implementation already works, it has been copied into this
version of NetworkController. The tests for `rollbackToPreviousProvider`
have also been updated to match the extension as well.