Refactor `lookupNetwork` test cases (#1318)
Refactor lookupNetwork test cases

The "lookup network" function is called internally in multiple methods,
and has a wide range of impacts. This has led to increasingly verbose
tests and inconsistencies in coverage.

All tests related to the lookup network step have been consolidated in
a helper function. This helper function is invoked separately for each
method that internally calls `lookupNetwork` (and for the tests that
call `lookupNetwork` directly). The one exception to this is
`rollbackToPreviousProvider`, which is not tested using this helper
function due to difficulties in abstracting the necessary test setup.

This relates to #1210