Refactor refresh network tests (#1335)
The network controller unit tests have been refactored to use a helper
function to test the effects of the refresh network step. This step
happens during a variety of different functions, and has the same
effects each time.

All tests related to the refresh network step have been consolidated in
a helper function. This helper function is invoked separately for each
method that internally calls `#refreshNetwork`. The one exception to
this is `rollbackToPreviousProvider`, which is not tested using this
helper function due to difficulties in abstracting the necessary test
setup.

This relates to #1210

---------

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>