Replace `NetworksChainId` constant with `ChainId`
The constant `NetworksChainId` was ambiguous in whether it referred to
network or chain Ids, and it didn't match our enum naming conventions.
It was also undocumented, and unclear in which networks it was intended
to contain.

It has been replaced with a `ChainId` constant with documentation and a
clear scope (built-in networks). A `BuiltInNetworks` enum has been
added as well to make that more clear.

Relates to #1209