Add (more) logging to NetworkController
NetworkController features asynchronous behavior that can be quite
tricky to debug. Specifically, for a brief moment in time, more than one
network may be "active" simultaneously. This can happen if the network
is switched before the code responsible for establishing a preconfigured
network (triggered via `initializeProvider` or one of the several ways
to switch the network) has yet to complete. This logging should aid us
in writing and debugging this behavior.