Preserve provider during lookupNetwork
If the network is switched while `lookupNetwork` is awaiting the request
for `net_version`, then the subsequent request for
`eth_getBlockByNumber` will take place against the new network, not the
original one. Despite the fact that this won't matter soon, as we will
shortly make it so that `lookupNetwork` won't make any state changes if
it detect a network change, it is quite confusing when debugging tests
that involve network switching in general (a future commit will
introduce logging so that such debugging is more possible). For this
reason, this commit changes `lookupNetwork` so that it stores a
reference to the provider (not the provider proxy) before doing anything
and passes that provider to the methods responsible for making requests.
This also means that we can't reuse `this.#ethQuery` in these methods,
as it relies on the provider proxy; we must make a new instance of
EthQuery before making the request. (This also raises a question about
the efficacy of the proxy, but that is a topic for another day.)