Consolidate state changes in lookupNetwork
Currently, `lookupNetwork` spawns two asynchronous operations. One
operation fetches the network ID and the other determines whether the
network supports EIP-1559. Because the second operation makes use of
`getEIP1559Compatibility`, it results in changing the controller state
implicitly. This is a bit strange, because `lookupNetwork` also changes
the controller state. In the extension version of `lookupNetwork`, the
steps are divided such that all state changes occur _after_ all
asynchronous operations complete. This commit brings the behavior of
`lookupNetwork` in line with the extension.