Support NFT detection on custom Mainnet RPC endpoints
The NFT detection controller now supports NFT detection on any Ethereum
Mainnet networks, including custom RPC endpoints. Previously this was
only supported on the built-in Ethereum Mainnet. This restriction was
likely not intentional in the first place (at least I can't think of a
reason for it).

This relates to #1209