Initialize asset controllers with the current network (#1361)
* Initialize asset controllers with the current network

Certain controllers in `@metamask/assets-controllers` would default to
assuming the current selected network was Ethereum Mainnet, or they'd
default with an empty string as the `chainId` (causing errors if any
methods were called). Instead they now all get initialized with the
current chain ID, passed in as a new constructor parameter.

Affected controllers are:
* `AssetsContractController`
* `NftController`
* `NftDetectionController`
* `TokenRatesController`
* `TokensController`

Relates to #1209

* Fix more tests