Replace hard-coded `chainId` on incoming token transactions
Incoming token transactions were given the correct chain ID on the
`txMeta` object, but the `transaction` object inside of that was given
a hard-coded chain ID of `1`. Likely we don't use this property for
anything, since we only keep track of these transactions to display
them as transaction history, and we'd use the `txMeta` property for
filtering those. But the type requires this property to be set, so it
should be correct.

Relates to #1209