Enable @typescript-eslint/no-explicit-any
There are many places across our packages where we type values with
`any`. While it is understandable that we might choose to "cheat" from
time to time due to difficult to use dependencies and time constraints —
and especially since this repository was originally a JavaScript
codebase — it creates technical debt that infects other code that uses
our libraries, so it is not desirable in the long run.

With that said, this commit corrects usages of `any` to either provide a
real type when we can or to use `// @ts-expect-error` to mark cases in
which the use `any` is intentional (for example, in tests that
purposefully pass bad input to the thing-under-test in order to verify
its behavior when used in a JavaScript context).