Remove `@metamask/types`
The `@metamask/types` package is no longer actively maintained, so it
has been removed. Most types we used to get from there were migrated to
`@metamask/utils` or can be gotten from `json-rpc-engine`.

Just one type had no replacement: `PermittedHandlerExport`. Since it is
specific to the permission controller anyway, it was added to that
package directly.