lookupNetwork: Drop mutex; bail on network change
If the network is changed multiple times in quick succession, multiple
"instances" of `lookupNetwork` can get run simultaneously. This happens
because each time the network is changed, `lookupNetwork` is called.
Because we do not expect `lookupNetwork` to run synchronously — it makes
HTTP requests and waits for them to complete — and because it makes
changes to the state, this creates a race condition where the state may
not reflect the last network that was selected.

Currently, to resolve this problem, `lookupNetwork` makes use of a
mutex. This ensures that only one "instance" of `lookupNetwork` is
running at a given time. However, this trades one problem for another,
because as a result, on mobile, the UI will be slow to update, as the
network controller has to "process" all of the network changes one at a
time.

So instead of using a mutex, a better solution is to copy the extension,
which makes it such if the network is changed while HTTP requests are
being awaited, no state changes will occur. This still guarantees that
later network changes override earlier ones while removing the
bottleneck.

Notable changes that I found while adding tests for this new behavior:

- When on an Infura network and the network is switched to a non-Infura
  network during a request for `net_version` or `eth_getBlockByNumber`,
  and Infura is blocking requests, `infuraIsBlocked` will no longer be
  emitted.
- Similarly, when on a non-Infura network and the network is switched to
  a Infura network during a request for `net_version` or
  `eth_getBlockByNumber`, and Infura is blocking requests,
  `infuraIsUnblocked` will no longer be emitted.