lookupNetwork: Drop mutex; bail on network change (#1375)
If the network is changed multiple times in quick succession, multiple
"instances" of `lookupNetwork` can get run simultaneously. This happens
because each time the network is changed, `lookupNetwork` is called.
Because we do not expect `lookupNetwork` to run synchronously — it makes
HTTP requests and waits for them to complete — and because it makes
changes to the state, this creates a race condition where the state may
not reflect the last network that was selected.

Currently, to resolve this problem, `lookupNetwork` makes use of a
mutex. This ensures that only one "instance" of `lookupNetwork` is
running at a given time. However, this trades one problem for another,
because as a result, on mobile, the UI will be slow to update, as the
network controller has to "process" all of the network changes one at a
time.

So instead of using a mutex, a better solution is to copy the extension,
which makes it such that if the network is changed while HTTP requests
are being awaited, no state changes will occur. This still guarantees that
later network changes override earlier ones while removing the
bottleneck.

Co-authored-by: Mark Stacey <markjstacey@gmail.com>