Update network status and details even when ID matches
The `lookupNetwork` method has been updated to ensure the network
status and details are updated correctly even when the network ID is
the same as before.

A test has been added to confirm that this works. It was failing before
this change.

When writing this test, I noticed that the previous test was broken as
well. It was passing but it didn't setup the test scenario correctly
(the network ID was changing, but it was expected not to). It has been
fixed as well.