Support hex and number `net_version` responses
Hex and number responses from the `net_version` request are now
accepted. While most chains return decimal strings for this request,
some chains were returning hex responses instead and failing our
validation for this request.

Support for number responses was added because it likely worked in
older versions of the extension  as well, so support is maintained to
avoid similar problems.

This is a port of the extension PR: https://github.com/MetaMask/metamask-extension/pull/19156