Add constraint to ensure controllers are peerDependencies (#1393)
* Add constraint to ensure controllers are peerDependencies

A constraint has been added to ensure that each controller that a
package depends on is also added as a `peerDependency`. This helps us
ensure that controllers appear only once in the dependency tree on the
client repositories.

Each controller is effectively a singleton because we only construct a
single instance of each one. Typically they get added as dependencies
only to get access to types in the controller package. If we don't add
it as a `peerDependency` as well, it risks the versions getting out-of
sync with what the client repository has installed.

Any missing `peerDependency` entries have been added to comply with
this constraint.

* Fix misleading local variable name

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>

---------

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>