Remove fake timers from NetworkController tests (#1399)
The extension version of the NetworkController tests do not use fake
timers. They were added for the tests in this repo because originally
the network client code was not being stubbed and so a real provider and
block tracker were being used (both of which use `setTimeout`
internally, to retry requests and progress the block tracker loop,
respectively). However, now that these tests are using a fake provider
and block tracker, we don't need to worry about using fake timers. This
change makes the tests more consistent with the extension version to
make the diff between them smaller.