Update getEIP1559Compatibility to match extension
There are two ways in which the `getEIP1559Compatibility` method in this
version of NetworkController differs from the same method in the
extension version:

- In this version, if `getEIP1559Compatibility` is called and
  `networkDetails.EIPS[1559]` in state is false, then the network is hit
  again in order to update the property. In the extension, however,
  that doesn't happen.
- In this version, if `getEIP1559Compatibility` is called but the
  provider has not been set yet, then the network is assumed to support
  EIP-1559 and the method returns true. In the extension, however,
  the network is assumed to _not_ support EIP-1559 and the method
  returns false.

In addition, the initial state of `networkDetails` in this version is `{
1559: false }`, whereas in the extension, it is `{}`. This is a problem
if `getEIP1559Compatibility` is updated to only hit the network if
`EIPS.1559` is unset, because it means the network will never be hit =
(at least initially).

This commit makes changes to align NetworkController with the extension
accordingly and updates the tests to match.